+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Add support for gasless transfers. Initially only supporting USDC sends on Base mainnet.
+- Add support for list historical balances for an asset of an address.
 
 ## [0.0.14] - 2024-08-05